update fedify example
