+++ examples/next15-app-router/app/[slug]/page.tsx
@@ -1,0 +1,8 @@
+export default async function Page({
+  params,
+}: {
+  params: Promise<{ slug: string }>;
+}) {
+  const slug = await params;
+  return <div>Hello {slug.slug}</div>;
+}
