+++ examples/next15-app-router/app/fedify-activity-handler/[[...catchAll]]/route.ts
@@ -1,0 +1,9 @@
+import { fedifyRequestHanlder } from "~/shared/integrate-fedify";
+
+export {
+  fedifyRequestHanlder as DELETE,
+  fedifyRequestHanlder as GET,
+  fedifyRequestHanlder as PATCH,
+  fedifyRequestHanlder as POST,
+  fedifyRequestHanlder as PUT,
+};
